Assign. 1.9 re-factor: add jQuery for crossing out element of list
Use .toggleClass() to add or remove one or more classes from each element in the set of matched elements, depending on either the class’s presence or the value of the state argument.
Use .on to attach an event handler function for one or more events to the selected elements.